+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 0.5.0 *(TBD)*
 ---------------------
 
 - NEW: `:options` command. open options page from omnibox.
+- NEW: `:version` command. now you can easily check omnitweety's version.
 
 
 Version 0.4.0 *(2015-12-03)*